rename package wave_generators to wavegenerators
