+++ zentone/src/main/java/com/github/nisrulz/zentone/wavegenerators/SquareWaveGenerator.kt
@@ -1,4 +1,4 @@
-package com.github.nisrulz.zentone.wave_generators
+package com.github.nisrulz.zentone.wavegenerators
 
 import kotlin.math.sign
 import kotlin.math.sin
