+++ zentone/src/main/java/com/github/nisrulz/zentone/wavegenerators/TriangleWaveGenerator.kt
@@ -1,4 +1,4 @@
-package com.github.nisrulz.zentone.wave_generators
+package com.github.nisrulz.zentone.wavegenerators
 
 import kotlin.math.asin
 import kotlin.math.sin
